requirements.txt
kafka-python
psycopg2
tweepy
vaderSentiment
